requirements.txt
blinker==1.9.0
boto3==1.35.91
botocore==1.35.91
CacheControl==0.14.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iso3166==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
mapbox==0.18.1
MarkupSafe==3.0.2
msgpack==1.1.0
polyline==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
